Handle private endpoint and misc
