+++ examples/src/main/java/dev/andstuff/kraken/example/JacksonTest.java
@@ -1,0 +1,34 @@
+package dev.andstuff.kraken.example;
+
+import java.util.Optional;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.MapperFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.json.JsonMapper;
+import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+
+public class JacksonTest {
+
+    private static final ObjectMapper OBJECT_MAPPER;
+
+    static {
+        OBJECT_MAPPER = JsonMapper.builder()
+                .enable(MapperFeature.ACCEPT_CASE_INSENSITIVE_ENUMS)
+                .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
+                .addModules(new JavaTimeModule(), new Jdk8Module())
+                .build();
+    }
+
+    public static void main(String[] args) throws JsonProcessingException {
+        MyRecord myRecord = OBJECT_MAPPER.readValue("{}", MyRecord.class);
+        System.out.println(myRecord);
+        System.out.println(myRecord.result().isEmpty());
+    }
+
+    public record MyRecord(Optional<MyClass> result) {}
+
+    public class MyClass {}
+}
