+++ library/src/main/java/dev/andstuff/kraken/api/model/KrakenException.java
@@ -1,0 +1,17 @@
+package dev.andstuff.kraken.api.model;
+
+import java.util.List;
+
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@ToString
+public class KrakenException extends RuntimeException {
+
+    private final List<String> errors;
+
+    public KrakenException(List<String> errors) {
+        this.errors = errors;
+    }
+}
