+++ library/src/main/java/dev/andstuff/kraken/api/model/endpoint/market/response/AssetStatus.java
@@ -1,0 +1,8 @@
+package dev.andstuff.kraken.api.model.endpoint.market.response;
+
+public enum AssetStatus {
+    ENABLED,
+    DEPOSIT_ONLY,
+    WITHDRAWAL_ONLY,
+    FUNDING_TEMPORARILY_DISABLED
+}
