+++ library/src/main/java/dev/andstuff/kraken/api/model/endpoint/market/response/ServerTime.java
@@ -1,0 +1,7 @@
+package dev.andstuff.kraken.api.model.endpoint.market.response;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record ServerTime(@JsonProperty("unixtime") long unixTime,
+                         String rfc1123) {
+}
