+++ library/src/main/java/dev/andstuff/kraken/api/model/endpoint/priv/PostParams.java
@@ -1,0 +1,8 @@
+package dev.andstuff.kraken.api.model.endpoint.priv;
+
+public interface PostParams { // TODO maybe create an abstract class
+
+    String initNonce();
+
+    String encoded();
+}
