+++ library/src/main/java/dev/andstuff/kraken/api/model/endpoint/pub/GenericQueryParams.java
@@ -1,0 +1,16 @@
+package dev.andstuff.kraken.api.model.endpoint.pub;
+
+import java.util.Map;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class GenericQueryParams implements QueryParams {
+
+    private final Map<String, String> params;
+
+    @Override
+    public Map<String, String> toMap() {
+        return params;
+    }
+}
