+++ library/src/main/java/dev/andstuff/kraken/api/model/endpoint/pub/QueryParams.java
@@ -1,0 +1,10 @@
+package dev.andstuff.kraken.api.model.endpoint.pub;
+
+import java.util.Map;
+
+public interface QueryParams {
+
+    QueryParams EMPTY = Map::of;
+
+    Map<String, String> toMap();
+}
