--- library/src/main/java/dev/andstuff/kraken/api/neo/model/KrakenCredentials.java
@@ -1,5 +1,0 @@
-package dev.andstuff.kraken.api.neo.model;
-
-public record KrakenCredentials(String key,
-                                String secret) {
-}
