--- library/src/main/java/dev/andstuff/kraken/api/neo/model/KrakenException.java
@@ -1,17 +1,0 @@
-package dev.andstuff.kraken.api.neo.model;
-
-import java.util.List;
-
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@ToString
-public class KrakenException extends RuntimeException {
-
-    private final List<String> errors;
-
-    public KrakenException(List<String> errors) {
-        this.errors = errors;
-    }
-}
