--- library/src/main/java/dev/andstuff/kraken/api/neo/model/endpoint/GenericPostParams.java
@@ -1,12 +1,0 @@
-package dev.andstuff.kraken.api.neo.model.endpoint;
-
-import java.util.Map;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class GenericPostParams implements PostParams {
-
-    private final Map<String, String> params;
-
-}
