--- library/src/main/java/dev/andstuff/kraken/api/neo/model/endpoint/JsonPrivateEndpoint.java
@@ -1,17 +1,0 @@
-package dev.andstuff.kraken.api.neo.model.endpoint;
-
-import java.util.Map;
-
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonNode;
-
-public class JsonPrivateEndpoint extends PrivateEndpoint<JsonNode> {
-
-    public JsonPrivateEndpoint(String path) {
-        this(path, Map.of());
-    }
-
-    public JsonPrivateEndpoint(String path, Map<String, String> postParams) {
-        super(path, new GenericPostParams(postParams), new TypeReference<>() {});
-    }
-}
