--- library/src/main/java/dev/andstuff/kraken/api/neo/model/endpoint/PostParams.java
@@ -1,4 +1,0 @@
-package dev.andstuff.kraken.api.neo.model.endpoint;
-
-public interface PostParams {
-}
