--- library/src/main/java/dev/andstuff/kraken/api/neo/model/endpoint/QueryParams.java
@@ -1,10 +1,0 @@
-package dev.andstuff.kraken.api.neo.model.endpoint;
-
-import java.util.Map;
-
-public interface QueryParams {
-
-    QueryParams EMPTY = Map::of;
-
-    Map<String, String> toMap();
-}
