--- library/src/main/java/dev/andstuff/kraken/api/neo/model/endpoint/market/ServerTimeEndpoint.java
@@ -1,12 +1,0 @@
-package dev.andstuff.kraken.api.neo.model.endpoint.market;
-
-import com.fasterxml.jackson.core.type.TypeReference;
-
-import dev.andstuff.kraken.api.neo.model.endpoint.market.response.ServerTime;
-
-public class ServerTimeEndpoint extends PublicEndpoint<ServerTime> {
-
-    public ServerTimeEndpoint() {
-        super("Time", new TypeReference<>() {});
-    }
-}
