--- library/src/main/java/dev/andstuff/kraken/api/neo/model/endpoint/market/params/GenericQueryParams.java
@@ -1,17 +1,0 @@
-package dev.andstuff.kraken.api.neo.model.endpoint.market.params;
-
-import java.util.Map;
-
-import dev.andstuff.kraken.api.neo.model.endpoint.QueryParams;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class GenericQueryParams implements QueryParams {
-
-    private final Map<String, String> params;
-
-    @Override
-    public Map<String, String> toMap() {
-        return params;
-    }
-}
