--- library/src/main/java/dev/andstuff/kraken/api/neo/model/endpoint/market/response/AssetStatus.java
@@ -1,8 +1,0 @@
-package dev.andstuff.kraken.api.neo.model.endpoint.market.response;
-
-public enum AssetStatus {
-    ENABLED,
-    DEPOSIT_ONLY,
-    WITHDRAWAL_ONLY,
-    FUNDING_TEMPORARILY_DISABLED
-}
