--- library/src/main/java/dev/andstuff/kraken/api/neo/model/endpoint/market/response/SystemStatus.java
@@ -1,7 +1,0 @@
-package dev.andstuff.kraken.api.neo.model.endpoint.market.response;
-
-import java.time.Instant;
-
-public record SystemStatus(String status, // TODO enum?
-                           Instant timestamp) {
-}
