--- library/src/main/java/dev/andstuff/kraken/api/neo/rest/KrakenRestRequester.java
@@ -1,8 +1,0 @@
-package dev.andstuff.kraken.api.neo.rest;
-
-import dev.andstuff.kraken.api.neo.model.endpoint.Endpoint;
-
-public interface KrakenRestRequester {
-
-    <T> T execute(Endpoint<T> endpoint);
-}
